requirements.txt
PyQt6>=6.7,<7.0
opencv-python>=4.10,<5.0
paddleocr>=2.8,<4.0
paddlepaddle>=3.3,<4.0
numpy>=1.24,<3.0
pyspellchecker>=0.8,<1.0
